refactor to bring yup schemas closer to the things they describe
