+++ client/constants/REPORT_PAGE_VALIDATION_ERRORS.json
@@ -1,0 +1,5 @@
+{
+  "MISSING_NUMBER": "Please enter a number.",
+  "MISSING_TIME_RANGE": "Please select a time range.",
+  "MISSING_SELECTION": "Plase make a selection."
+}
